+++ src/codecs/codecs.jl
@@ -8,17 +8,21 @@ would require, and NOT the actual encoding/decoding function itself.
 This makes things customizable, and allows plug-and-play behaviour with
 different variants of the same codec scheme, or for strikingly similar codec
 schemes.
-
-!!! note
-    Every Codec object MUST have exactly one set of `encode_f::Function` and
-    `decode_f::Function` fields.
 """
 module Codec
+
+using ..ChemistryFeaturization.AbstractType: AbstractCodec
+import ..ChemistryFeaturization.output_shape
 
 include("simplecodec.jl")
 export SimpleCodec
 
 include("onehotonecold.jl")
-export OneHotOneCold, build_onehot_vec
+export OneHotOneCold, output_shape
+
+include("directcodec.jl")
+export DirectCodec
+
+export encode, decode
 
 end