+++ test/codecs/codec_tests.jl
@@ -1,0 +1,9 @@
+using ChemistryFeaturization.Codec
+
+@testset "Codecs" begin
+    codec_tests = ["OneHotOneCold_tests", "SimpleCodec_tests", "DirectCodec_tests"]
+    for t in codec_tests
+        tp = abspath(testdir, "codecs", "$(t).jl")
+        include(tp)
+    end
+end